# api/public/schema/openapi.yaml
openapi: 3.0.2
info:
  title: 'ID Vault'
  version: V1.0
  description: |
    API Details
    - Component: ID Vault
    - Reference: id-vault
    - Enviroment: dev
    - Version: V1.0
    - Repository: [online](https://github.com/ConductionNL/id-vault) / [zip](https://github.com/ConductionNL/id-vault/archive/master.zip)
    - Docker Image: [online](https://hub.docker.com/repository/docker/conduction/id-vault-php) 
    - Datamodel: [postman](/schema/id-vault.postman_collection.json) / [oas](/schema/openapi.yaml) / [pdf](/schema/datamodel.pdf) / [modelio (zip)](/schema/Datamodel_Modelio.zip) 
    
    Naast deze JSON rest API is er ook een [graphql](/graphql) interface beschikbaar.         
paths:
...
  /api/send_lists:
    get:
      tags:
        - SendList
      operationId: getSendListCollection
      summary: 'Retrieves the collection of SendList resources.'
      responses:
        200:
          description: 'SendList collection response'
          content:
            application/ld+json:
              schema:
                type: object
                properties:
                  'hydra:member': { type: array, items: { $ref: '#/components/schemas/SendList:jsonld-read' } }
                  'hydra:totalItems': { type: integer, minimum: 0 }
                  'hydra:view': { type: object, properties: { '@id': { type: string, format: iri-reference }, '@type': { type: string }, 'hydra:first': { type: string, format: iri-reference }, 'hydra:last': { type: string, format: iri-reference }, 'hydra:next': { type: string, format: iri-reference } } }
                  'hydra:search': { type: object, properties: { '@type': { type: string }, 'hydra:template': { type: string }, 'hydra:variableRepresentation': { type: string }, 'hydra:mapping': { type: array, items: { type: object, properties: { '@type': { type: string }, variable: { type: string }, property: { type: string }, required: { type: boolean } } } } } }
                required:
                  - 'hydra:member'
            application/json+ld:
              schema:
                type: object
                properties:
                  'hydra:member': { type: array, items: { $ref: '#/components/schemas/SendList:jsonld-read' } }
                  'hydra:totalItems': { type: integer, minimum: 0 }
                  'hydra:view': { type: object, properties: { '@id': { type: string, format: iri-reference }, '@type': { type: string }, 'hydra:first': { type: string, format: iri-reference }, 'hydra:last': { type: string, format: iri-reference }, 'hydra:next': { type: string, format: iri-reference } } }
                  'hydra:search': { type: object, properties: { '@type': { type: string }, 'hydra:template': { type: string }, 'hydra:variableRepresentation': { type: string }, 'hydra:mapping': { type: array, items: { type: object, properties: { '@type': { type: string }, variable: { type: string }, property: { type: string }, required: { type: boolean } } } } } }
                required:
                  - 'hydra:member'
            application/hal+json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SendList-read'
            application/json+hal:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SendList-read'
            application/vnd.api+json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SendList-read'
            application/json+vnd.api:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SendList-read'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SendList-read'
            application/health+json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SendList-read'
            application/json+health:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SendList-read'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SendList-read'
            application/x-yaml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SendList-read'
            text/csv:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SendList-read'
            text/html:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SendList-read'
      parameters:
        -
          name: 'order[id]'
          in: query
          required: false
          schema:
            type: string
            enum:
              - asc
              - desc
        -
          name: 'order[dateCreated]'
          in: query
          required: false
          schema:
            type: string
            enum:
              - asc
              - desc
        -
          name: 'order[dateModified]'
          in: query
          required: false
          schema:
            type: string
            enum:
              - asc
              - desc
        -
          name: 'dateCreated[before]'
          in: query
          required: false
          schema:
            type: string
        -
          name: 'dateCreated[strictly_before]'
          in: query
          required: false
          schema:
            type: string
        -
          name: 'dateCreated[after]'
          in: query
          required: false
          schema:
            type: string
        -
          name: 'dateCreated[strictly_after]'
          in: query
          required: false
          schema:
            type: string
        -
          name: 'dateModified[before]'
          in: query
          required: false
          schema:
            type: string
        -
          name: 'dateModified[strictly_before]'
          in: query
          required: false
          schema:
            type: string
        -
          name: 'dateModified[after]'
          in: query
          required: false
          schema:
            type: string
        -
          name: 'dateModified[strictly_after]'
          in: query
          required: false
          schema:
            type: string
        -
          name: id
          in: query
          required: false
          schema:
            type: string
        -
          name: 'id[]'
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        -
          name: dateCreated
          in: query
          required: false
          schema:
            type: string
        -
          name: 'dateCreated[]'
          in: query
          required: false
          schema:
            type: string
        -
          name: dateModified
          in: query
          required: false
          schema:
            type: string
        -
          name: 'dateModified[]'
          in: query
          required: false
          schema:
            type: string
        -
          name: page
          in: query
          required: false
          description: 'The collection page number'
          schema:
            type: integer
            default: 1
        -
          name: limit
          in: query
          required: false
          description: 'The number of items per page'
          schema:
            type: integer
            default: 30
            minimum: 0
        -
          name: Authorization
          description: 'The JWT of the entity performing the request'
          in: header
        -
          name: API-Version
          description: 'The version of the API conform [Landelijke API-strategie.](https://geonovum.github.io/KP-APIs/#versioning)'
          example: 1.0.1
          in: header
        -
          name: X-NLX-Logrecord-ID
          description: 'A  globally unique id of the request, which makes a request traceable throughout the network.'
          in: header
        -
          name: X-NLX-Request-Process-Id
          description: 'A process id for purpose registration (doelbinding)'
          in: header
        -
          name: X-NLX-Request-Data-Elements
          description: 'A list of requested data elements'
          in: header
        -
          name: X-NLX-Request-Data-Subject
          description: 'A key-value list of data subjects related to this request. e.g. `bsn=12345678, kenteken=ab-12-fg`'
          in: header
        -
          name: X-NLX-Audit-Clarification
          description: 'A clarification as to why a request has been made  (doelbinding)'
          in: header
        -
          name: Link
          description: 'A [websub](https://www.w3.org/TR/websub/#discovery) header like <https://hub.example.com/>; rel="hub"'
          in: header
        -
          name: 'extend[]'
          required: false
          description: 'An array of nested objects to include in the return object'
          in: query
          schema:
            type: array
        -
          name: 'fields[]'
          required: false
          description: 'An array of fields to return in output, wil return all fields is not supplied'
          in: query
          schema:
            type: array
        -
          name: validOn
          required: false
          description: 'Returns object as valid on a given date time'
          schema:
            type: string
            format: date-time
          in: query
        -
          name: validFrom
          required: false
          description: 'Returns objects valid from a given date time'
          schema:
            type: string
            format: date-time
          in: query
        -
          name: validUntil
          required: false
          description: 'Returns objects valid until a given date time'
          schema:
            type: string
            format: date-time
          in: query
      produces:
        - application/health+json
    post:
      tags:
        - SendList
      operationId: postSendListCollection
      summary: 'Creates a SendList resource.'
      responses:
        201:
          description: 'SendList resource created'
          content:
            application/ld+json:
              schema:
                $ref: '#/components/schemas/SendList:jsonld-read'
            application/json+ld:
              schema:
                $ref: '#/components/schemas/SendList:jsonld-read'
            application/hal+json:
              schema:
                $ref: '#/components/schemas/SendList-read'
            application/json+hal:
              schema:
                $ref: '#/components/schemas/SendList-read'
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/SendList-read'
            application/json+vnd.api:
              schema:
                $ref: '#/components/schemas/SendList-read'
            application/json:
              schema:
                $ref: '#/components/schemas/SendList-read'
            application/health+json:
              schema:
                $ref: '#/components/schemas/SendList-read'
            application/json+health:
              schema:
                $ref: '#/components/schemas/SendList-read'
            application/xml:
              schema:
                $ref: '#/components/schemas/SendList-read'
            application/x-yaml:
              schema:
                $ref: '#/components/schemas/SendList-read'
            text/csv:
              schema:
                $ref: '#/components/schemas/SendList-read'
            text/html:
              schema:
                $ref: '#/components/schemas/SendList-read'
          links:
            GetSendListItem:
              parameters:
                id: '$response.body#/id'
              operationId: getSendListItem
              description: 'The `id` value returned in the response can be used as the `id` parameter in `GET /api/send_lists/{id}`.'
        400:
          description: 'Invalid input'
        404:
          description: 'Resource not found'
      requestBody:
        content:
          application/ld+json:
            schema:
              $ref: '#/components/schemas/SendList:jsonld-write'
          application/json+ld:
            schema:
              $ref: '#/components/schemas/SendList:jsonld-write'
          application/hal+json:
            schema:
              $ref: '#/components/schemas/SendList-write'
          application/json+hal:
            schema:
              $ref: '#/components/schemas/SendList-write'
          application/vnd.api+json:
            schema:
              $ref: '#/components/schemas/SendList-write'
          application/json+vnd.api:
            schema:
              $ref: '#/components/schemas/SendList-write'
          application/json:
            schema:
              $ref: '#/components/schemas/SendList-write'
          application/health+json:
            schema:
              $ref: '#/components/schemas/SendList-write'
          application/json+health:
            schema:
              $ref: '#/components/schemas/SendList-write'
          application/xml:
            schema:
              $ref: '#/components/schemas/SendList-write'
          application/x-yaml:
            schema:
              $ref: '#/components/schemas/SendList-write'
          text/csv:
            schema:
              $ref: '#/components/schemas/SendList-write'
          text/html:
            schema:
              $ref: '#/components/schemas/SendList-write'
        description: 'The new SendList resource'
      parameters:
        -
          name: Authorization
          description: 'The JWT of the entity performing the request'
          in: header
        -
          name: API-Version
          description: 'The version of the API conform [Landelijke API-strategie.](https://geonovum.github.io/KP-APIs/#versioning)'
          example: 1.0.1
          in: header
        -
          name: X-NLX-Logrecord-ID
          description: 'A  globally unique id of the request, which makes a request traceable throughout the network.'
          in: header
        -
          name: X-NLX-Request-Process-Id
          description: 'A process id for purpose registration (doelbinding)'
          in: header
        -
          name: X-NLX-Request-Data-Elements
          description: 'A list of requested data elements'
          in: header
        -
          name: X-NLX-Request-Data-Subject
          description: 'A key-value list of data subjects related to this request. e.g. `bsn=12345678, kenteken=ab-12-fg`'
          in: header
        -
          name: X-NLX-Audit-Clarification
          description: 'A clarification as to why a request has been made  (doelbinding)'
          in: header
...
components:
  schemas:
...
    SendList-read:
      type: object
      description: ''
      properties:
        id:
          readOnly: true
          description: 'The UUID identifier of this resource'
          externalDocs:
            url: 'http://schema.org/identifier'
          type: string
          format: uuid
        resource:
          description: 'A BS/SendList resource. Used for Adding a user as BS/Subscriber to a BS/SendList. And used for sending an email to all BS/SendList->Subscribers.'
          externalDocs:
            url: 'http://schema.org/url'
          type: string
        name:
          description: 'The name of a new SendList. Used for creating a BS/SendList.'
          type: string
        description:
          description: 'The description of a new SendList. Used for creating a BS/SendList.'
          type: string
        mail:
          description: 'True if this is a new mailing sendList. Used for creating a BS/SendList.'
          type: boolean
        phone:
          description: 'True if this is a new phone sendList. Used for creating a BS/SendList.'
          type: boolean
        title:
          description: 'The title for sending an text/email to all BS/SendList->Subscribers.'
          type: string
        message:
          description: 'The message for sending an email to all BS/SendList->Subscribers.'
          type: string
        text:
          description: 'The text for sending an text to all BS/SendList->Subscribers.'
          type: string
        html:
          description: 'The html for sending an email to all BS/SendList->Subscribers.'
          type: string
        result:
          description: 'The result'
          type: array
          items:
            type: string
        dateCreated:
          description: 'The moment this claim was created'
          type: string
          format: date-time
          nullable: true
        dateModified:
          description: 'The moment this claim was last Modified'
          type: string
          format: date-time
          nullable: true
    SendList-write:
      type: object
      description: ''
      properties:
        resource:
          description: 'A BS/SendList resource. Used for Adding a user as BS/Subscriber to a BS/SendList. And used for sending an email to all BS/SendList->Subscribers.'
          externalDocs:
            url: 'http://schema.org/url'
          type: string
        name:
          description: 'The name of a new SendList. Used for creating a BS/SendList.'
          type: string
        description:
          description: 'The description of a new SendList. Used for creating a BS/SendList.'
          type: string
        mail:
          description: 'True if this is a new mailing sendList. Used for creating a BS/SendList.'
          type: boolean
        phone:
          description: 'True if this is a new phone sendList. Used for creating a BS/SendList.'
          type: boolean
        clientSecret:
          description: 'The secret of your application. Used for creating a BS/SendList.'
          type: string
        title:
          description: 'The title for sending an text/email to all BS/SendList->Subscribers.'
          type: string
        message:
          description: 'The message for sending an email to all BS/SendList->Subscribers.'
          type: string
        text:
          description: 'The text for sending an text to all BS/SendList->Subscribers.'
          type: string
        html:
          description: 'The html for sending an email to all BS/SendList->Subscribers.'
          type: string
    'SendList:jsonld-read':
      type: object
      description: ''
      properties:
        '@context':
          readOnly: true
          type: string
        '@id':
          readOnly: true
          type: string
        '@type':
          readOnly: true
          type: string
        id:
          readOnly: true
          description: 'The UUID identifier of this resource'
          externalDocs:
            url: 'http://schema.org/identifier'
          type: string
          format: uuid
        resource:
          description: 'A BS/SendList resource. Used for Adding a user as BS/Subscriber to a BS/SendList. And used for sending an email to all BS/SendList->Subscribers.'
          externalDocs:
            url: 'http://schema.org/url'
          type: string
        name:
          description: 'The name of a new SendList. Used for creating a BS/SendList.'
          type: string
        description:
          description: 'The description of a new SendList. Used for creating a BS/SendList.'
          type: string
        mail:
          description: 'True if this is a new mailing sendList. Used for creating a BS/SendList.'
          type: boolean
        phone:
          description: 'True if this is a new phone sendList. Used for creating a BS/SendList.'
          type: boolean
        title:
          description: 'The title for sending an text/email to all BS/SendList->Subscribers.'
          type: string
        message:
          description: 'The message for sending an email to all BS/SendList->Subscribers.'
          type: string
        text:
          description: 'The text for sending an text to all BS/SendList->Subscribers.'
          type: string
        html:
          description: 'The html for sending an email to all BS/SendList->Subscribers.'
          type: string
        result:
          description: 'The result'
          type: array
          items:
            type: string
        dateCreated:
          description: 'The moment this claim was created'
          type: string
          format: date-time
          nullable: true
        dateModified:
          description: 'The moment this claim was last Modified'
          type: string
          format: date-time
          nullable: true
    'SendList:jsonld-write':
      type: object
      description: ''
      properties:
        '@context':
          readOnly: true
          type: string
        '@id':
          readOnly: true
          type: string
        '@type':
          readOnly: true
          type: string
        resource:
          description: 'A BS/SendList resource. Used for Adding a user as BS/Subscriber to a BS/SendList. And used for sending an email to all BS/SendList->Subscribers.'
          externalDocs:
            url: 'http://schema.org/url'
          type: string
        name:
          description: 'The name of a new SendList. Used for creating a BS/SendList.'
          type: string
        description:
          description: 'The description of a new SendList. Used for creating a BS/SendList.'
          type: string
        mail:
          description: 'True if this is a new mailing sendList. Used for creating a BS/SendList.'
          type: boolean
        phone:
          description: 'True if this is a new phone sendList. Used for creating a BS/SendList.'
          type: boolean
        clientSecret:
          description: 'The secret of your application. Used for creating a BS/SendList.'
          type: string
        title:
          description: 'The title for sending an text/email to all BS/SendList->Subscribers.'
          type: string
        message:
          description: 'The message for sending an email to all BS/SendList->Subscribers.'
          type: string
        text:
          description: 'The text for sending an text to all BS/SendList->Subscribers.'
          type: string
        html:
          description: 'The html for sending an email to all BS/SendList->Subscribers.'
          type: string
...
securityDefinitions:
  JWT-Oauth:
    type: oauth2
    authorizationUrl: 'http://petstore.swagger.io/api/oauth/dialog'
    flow: implicit
    scopes:
      read: 'read right to the ChangeLog resource'
      write: 'write right to the ChangeLog resource'
  JWT-Token:
    type: apiKey
    in: header
    name: Authorization
    scopes:
      read: 'read right to the ChangeLog resource'
      write: 'write right to the ChangeLog resource'
host: irc.zaakonline.nl
servers:
  -
    name: zuid-drecht.nl
    url: id-vault.dev.zuid-drecht.nl
...